🚚 Rename and move solo stats widget
