+++ src/lib/Widgets/SoloStats.svelte
